Revert "Move `String.toUrl` into test"
This reverts commit 9968e01f116019863256d09d32602b592f911372.
